requirements.txt
torch>=1.8.0
torchvision>=0.9.0
cn_clip>=1.5
timm>=0.4.12
numpy>=1.21.0
psycopg2-binary>=2.9.0
Pillow>=9.0.0